NEW/MVP/mobile-server/requirements.txt
# Tunnel Proxy
websockets>=12.0
httpx>=0.25.0
aiohttp>=3.9.0
pyyaml>=6.0

# Telegram (Pyrogram)
pyrogram>=2.0.0
tgcrypto>=1.2.0

# Web frameworks
fastapi>=0.104.0
uvicorn>=0.24.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0

# Browser automation (optional)
# playwright>=1.40.0
